Update README.md
Add in a notice about the lack of maintenance on this repository. Sadly we haven't got the people to maintain this repository and we'd like to make it clear to visitors to this repository about it's state. This process of marking it unmaintained only affects future updates. The Action will still continue work as it is in old workflows.

Once this lands, we'll also Archive this repository.